Move components to their own file
